requirements.txt
django 
djangorestframework
django-cors-headers
django-allauth
pymysql
django-csp<3.7.0
